Reorganizing feature tests for homepage, search page and item page.
