--- spec/models/agreement_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-describe Agreement do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
